Fix usage of 'getParam()'
The `getParam()`-function accepts a `default` parameter to return a default
value in case no parameter with the corresponding key could be found.
The function did only check for the truthiness of the `default` parameter
thus when passed `0`, `{}` or `False` as default, the function would return
`None`.
This commit fixes this behaviour.